fix(meai-10-5): review-A — IMcpToolSource port fakes + internal Configuration accessor
- HIGH: replace 7 NSubstitute mocks of IMcpToolSource with InProcessTestMcpToolSource (extracted to Fixtures/)
- MED: expose internal AgentStepBase.Configuration; drop reflection helper in AgentStepBuilderToolsTests
- LOW: seal AgentDiagnosticsTests

Refs: PR #82 review verdict
